% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWCCalculator-package.R
\docType{package}
\name{FWCCalculator-package}
\alias{FWCCalculator}
\alias{FWCCalculator-package}
\title{FWCCalculator: Calculate Flow Weights and Event Mean Concentrations from Water Quality Data}
\description{
A Shiny app for calculating the flow-weighting and event mean concentration of water quality data with discrete samples.
}
\author{
\strong{Maintainer}: Nicholas Lombardo \email{nicholasl@sccwrp.org}

Other contributors:
\itemize{
  \item Edward Tiernan \email{edwardt@sccwrp.org} [contributor]
}

}
\keyword{internal}
