% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWCCalculator-package.R
\docType{package}
\name{FWCCalculator-package}
\alias{FWCCalculator}
\alias{FWCCalculator-package}
\title{FWCCalculator: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
